refactor(frontend): split the developer portal pages into focused components
Addresses the two Aikido long_function findings. Both pages held their data
fetching, every piece of presentation state, and the whole render in one
function, so following any single behaviour meant scanning past unrelated logic.

DeveloperApiKeys 278 -> 133 lines, with three components beside it:

  CreateKeyForm  owns the name/environment/scopes fields. The page unmounts it
                 on success, which is what clears them - so resetForm is gone,
                 and a failed create keeps what was typed because the form
                 stays mounted.
  KeyReveal      owns `copied`, which is only meaningful while the plaintext key
                 is on screen. Unmounting on Done clears it.
  KeyTable       presentational. Loading and empty are passed in rather than
                 inferred, since keys.length === 0 cannot tell "none yet" from
                 "not loaded yet".

DeveloperBilling 358 -> 146 lines, likewise:

  CurrentPlanRow the summary row and its status badge, plus formatPeriod.
  PlanCard       one tier, plus the PLANS copy it renders and the proLabel
                 the three button states need.
  UsageMeter     moved out whole, unchanged.

What is left in each page is the server-backed state and the requests: the key
list and its three operations, and the two reads plus the two Stripe hand-offs.

No behaviour change, and the existing tests are the evidence - all 104 developer
tests pass untouched. Every extracted file is at 100% statements except PlanCard
at 99.35%. Re-ran the browser QA over a fresh Storybook build: 11 stories, zero
axe violations, screenshots unchanged.
